feat: add HeirloomEntity interface, ResourceType implements it
- New HeirloomEntity interface with mandatory + default-null optional fields
- ResourceType now implements HeirloomEntity
- Added fields: fullyQualifiedName, domain, changeHash, deleted
- Added getEntityType() returning 'resourceType'
- TypeResponse.version changed from int to Long
- All 13 existing tests pass

+++ heirloom-server/src/main/java/com/heirloom/schema/dto/TypeResponse.java
@@ -15,7 +15,7 @@ public record TypeResponse(
         List<Ability> abilities,
         List<StateTransition> stateMachine,
         List<Relationship> relationships,
-        int version,
+        Long version,
         Instant createdAt,
         Instant updatedAt
 ) {
